+++ src/components/common/Navigator/TopNavBar/Elements/ChatAvatars.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Avatar } from "../..";
+import { Avatar } from "../../..";
 import { styled } from "styled-components";
 
 interface IChatAvatarsProps {
