+++ src/components/common/Navigator/TopNavBar/Elements/LogoImage.tsx
